requirements.txt
pytest
pyside2
numpy
matplotlib
pybluez
leap-protocol